Eliminação do pacote pojo
